refactor(nuget): Separate `resolveDependencies` out of the inspector
The `NuGetInspector` should just be a wrapper around the CLI tool (with
capabilities to transform its results), but the actual dependency
resolution / building of the ORT data model should be the concern of the
`NuGet` implementation.

Note that further changes to fixup the error / warning mapping will
follow.
